Initial refactoring of codebase
Started refactoring the codebase. Moved the Settings out to a different
project so it's accessible by the Geonet stuff. All quakes are now
accessed from a singleton EarthquakeContainer which will provide you
with the right number of quakes as per the settings, classes just add
themselves as a listener.

Currently the single quake display page is broken.

Still a lot of code (and duplication) to refactor out.