Added connectivity issue toast messages
For when Geonet sends bad data, or no network, etc. Haven't tested that
this works because I don't have a device to test on... and I can't
disable network connectiving on the emulator. gg. Haven't got around to
figuring out how to test GUI stuff via the test classes yet; when I do,
I should be able to use the mocks for faking bad connections and make
sure these work.