Updated tests to work
I don't understand why a) the GeoCoordinate.Equals() method no longer
seems to work as expected (I would've thought floating point comparisons
were to blame, but comparing the Latitude and Longitude values directly
works as expected...) or b) any direct reference to a JsonException
causes the tests running on the emulator to crash with a "network
connection has been lost" message in Visual Studio and no other
information. This happens with both the old way by calling typeof() and
GetType(), or the way I've implemented in this commit (but commented
out) of catching the JsonException and calling Assert.Fail() on any
other type of exception.