Deleted splash screen image, updated icon
Also updated gitignore properly this time, apparently comments on the
same line don't actually work. Added a 300x300 icon for the phone
application list (since I can't seem to find the required size, this one
gets downscaled so that's better than too-small) since the one I was
using was off-center. Added a 300x300 non-transparent icon as well, for
the marketplace.